requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
python-dotenv
python-jose
stripe